--- a/build.sbt
+++ b/build.sbt
@@ -14,3 +14,18 @@ scalacOptions := Seq(
 licenses += ("Apache-2.0", url("https://www.apache.org/licenses/LICENSE-2.0.html"))
 
 scalafmtOnCompile := true
+
+libraryDependencies += {
+  val utestVersion = scalaBinaryVersion.value match {
+    case "2.10" | "2.11" => "0.6.8"
+    case _ => "0.7.4"
+  }
+  "com.lihaoyi" %% "utest" % utestVersion % Test
+}
+testFrameworks += new TestFramework("utest.runner.Framework")
+
+enablePlugins(SbtPlugin)
+scriptedLaunchOpts := {
+  scriptedLaunchOpts.value ++
+    Seq("-Xmx1024M", "-Dplugin.version=" + version.value)
+}