--- a/src/main/scala/com/github/sbt/duplicates/Checksum.scala
+++ b/src/main/scala/com/github/sbt/duplicates/Checksum.scala
@@ -1,5 +1,0 @@
-package com.github.sbt.duplicates
-
-import sbt.File
-
-case class Checksum(source: File, checksum: String)