--- a/src/sbt-test/sanity/sanity/differ/src/main/scala/ConflictingFile.scala
+++ b/src/sbt-test/sanity/sanity/differ/src/main/scala/ConflictingFile.scala
@@ -1,0 +1,4 @@
+class ConflictingFile {
+
+  def different = true
+}