--- a/src/sbt-test/sanity/sanity/equal/src/main/scala/ConflictingFile.scala
+++ b/src/sbt-test/sanity/sanity/equal/src/main/scala/ConflictingFile.scala
@@ -1,0 +1,4 @@
+class ConflictingFile {
+
+  def core = true
+}