--- a/src/sbt-test/sanity/sanity/project/plugins.sbt
+++ b/src/sbt-test/sanity/sanity/project/plugins.sbt
@@ -1,0 +1,9 @@
+{
+  val pluginVersion = System.getProperty("plugin.version")
+  if (pluginVersion == null)
+    throw new RuntimeException(
+      """|The system property 'plugin.version' is not defined.
+         |Specify this property using the scriptedLaunchOpts -D.""".stripMargin
+    )
+  else addSbtPlugin("com.github.sbt" % "sbt-duplicates-finder" % pluginVersion)
+}