--- a/src/test/scala-sbt-1.0/com/github/sbt/duplicates/TestCompat.scala
+++ b/src/test/scala-sbt-1.0/com/github/sbt/duplicates/TestCompat.scala
@@ -1,0 +1,5 @@
+package com.github.sbt.duplicates
+
+object TestCompat {
+  val SbtIO = sbt.io.IO
+}