## Summary
Decreases `checkDuplicatesTest` runtime from 150s to 10s by avoiding unnecessary filesystem accesses and hashing.

Added a happy-path unit test for the things I changed.

## Problem
`checkDuplicatesTest` takes roughly 150 seconds to complete on my 27-project build.

CPU usage is quite high:
![checkDuplicatesTest uses maxes out all 12 cpu cores](https://user-images.githubusercontent.com/663139/87760303-89186a80-c7dd-11ea-8325-00ce5d55cdb0.png)

## Other Things I Tried

### Attempt 1: SHA-256 => Murmur
Initially, I suspected the SHA-256 hashing. What else could be burning through so much CPU time?
![1 0 0 hashing](https://user-images.githubusercontent.com/663139/87760693-2378ae00-c7de-11ea-8035-d20fd0f19848.png)

I tried switching to scala's built in Murmur hash function, but it only cut the time down to ~110 seconds.

### Attempt 2: Caching
I figured with 27 projects, there had to be some overlap in my classpath. I dumped `fullClasspath` in each of my projects then flattened:

```
cat **/fullClasspath.txt > all.txt
$ wc -l all.txt
    5495 all.txt
$ cat all.txt | sort | uniq | wc -l
     533
```

So `checkDuplicatesTest` is opening and rehashing each file ~10x. If we had some way to cache the results, we could eliminate this work. But how?

I messed around with a `ConcurrentHashMap`, but couldn't find a good signal to purge it to avoid leaking memory across repeated runs.

I read through [sbt Caching](https://www.scala-sbt.org/1.x/docs/Caching.html), but a few things turned me away.
1. Refactoring - seems like it wants to live at the task level rather than the pure scala level. I didn't want to refactor too much.
2. Filesystem load - looks like it writes to the filesystem. One write per class file seems heavy.

Ultimately, I abandoned the caching approach, since it seems non-trivial to do well.

### Attempt 3: Avoid FS access & hashing
Thinking about FS load, I realized that half of the CPU time is spent in system calls. Is this FS activity?
![1 0 0 system](https://user-images.githubusercontent.com/663139/87760661-1360ce80-c7de-11ea-8f31-43f6d8812f20.png)

Looking deeper, I realized that we're reading and hashing everything, but we only need the hashes if there are multiple files with the same FQCN. Many of these names are unique. We don't need to read or hash these files at all!

I refactored to defer reading the files until after we check if the FQCN is duplicated. I expect that this change was responsible for most of the speedup.

I also spotted that we have a perfectly good CRC32 pre-calculated in the jar/zip central directory. We can grab it for free while traversing the zip entries for the names.

For classes in directories (i.e. outside of jars), we have to compute the CRC32 ourselves, but only if needed on FQCN conflict. I expect the ratio of jars to class files is probably high.

There's some debate on https://stackoverflow.com/q/10953958 about how good CRC32 is as a hash function. (I assume CRC32 and CRC32C perform similarly). It's clearly not as strong as SHA-256, but it also doesn't have to be. On collision, we can just read the whole file contents. Good enough.

### Attempt 4: Back to SHA-256
Thinking further, the CRC32 comparison adds less value than I thought. It only saves IO for the ConflictState.ContentDiffer case. That's the exceptional case rather than the norm since we're probably running checkDuplicatesTest in a CI build. Over the common case, the crc32s will collide and we'll have to go to disk anyway.

checksums.combinations(2).map{ c1 == c2 } perf could regress pathologically badly after my changes in a scenario with lots of duplicates with equal content.

We could replace the combinations with a `checksums.groupBy(_.bytes)` to get back to O(n) disk reads. Simple, but not memory-safe. To get memory safety back, we could either go back to a wide SHA (simple and probabilistically correct) or stream the content and compare bytewise (more complex, faster, 100% correct).

I ended up tossing the CRC32 and bytewise comparison and added back the original SHA-256. Seems like a good balance.

## CLA
I don't see Lightbend CLA checks here, but FWIW, I've signed it previously. You can verify against my akka/akka PRs.