@pdalpra 2020-07-17
Thanks for this very detailed PR, the work you've put into it is impressive :astonished: 

I'll have a deeper look ASAP, and handle the few upgrades that did not work (as you can see, there has been little maintenance from me on this plugin...)

> I don't see Lightbend CLA checks here, but FWIW, I've signed it previously. You can verify against my akka/akka PRs.

There is no need for the CLA is this case, the plugin is hosted in the `sbt` organization but is not a Lightbend product.