@htmldoug 2020-07-17
Thanks for the quick response! I've rebased onto master and fixed the first CI failure. Looks like `sbt.io` moved around. I'll fix that next.

> there has been little maintenance from me on this plugin

Hey, it's simple and does the job. Not much is needed. :)