@pdalpra 2020-07-17
FYI, I just updated sbt to 1.3.13 in #76.