@htmldoug 2020-07-17
Thanks for the heads up. Should be all rebased and passing. I'm done force pushing for now and going to bed. Feel free to change anything to your liking.