@htmldoug 2020-07-19
Ready for review. I'm done for now. Should be in good shape. Happy to make any changes.