@htmldoug 2020-07-19
I think I've addressed all comments now. Thanks for the thoughtful review and for teaching me some new sbt protips.