@pdalpra 2020-07-21
Just released sbt-duplicates-finder 1.1.0, it should be synced to the sbt mirror soon (you can check when it becomes available here: https://bintray.com/sbt/sbt-plugin-releases)